+++ hc3d-tm/src/business_logic/temp_validator.h
@@ -11,8 +11,8 @@
 #include "stdint.h"
 #include "stdbool.h"
 
-extern uint16_t temp_validator_sensor_last_valid_temperature[HC3D_CONFIG_TEMP_SENSOR_COUNT];
-extern uint32_t temp_validator_sensor_last_update_time[HC3D_CONFIG_TEMP_SENSOR_COUNT];
+extern uint8_t temp_validator_sensor_last_valid_temperature[HC3D_CONFIG_TEMP_SENSOR_COUNT];
+extern uint16_t temp_validator_sensor_last_update_time[HC3D_CONFIG_TEMP_SENSOR_COUNT];
 
 void temp_validator_init(void);
 void temp_validator_tick(void);
